geolocation/geo-00001.html

<!DOCTYPE html>
<html>
  <head>
    <title>Geolocation Test: error callback is called with correct code, when user denies access</title>
    <link rel="author" title="Intel" href="http://www.intel.com/">
    <link rel="help" href="http://www.w3.org/TR/2012/PR-geolocation-API-20120510/#geolocation">
    <link rel="help" href="http://www.w3.org/TR/2012/PR-geolocation-API-20120510/#get-current-position">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      var t =async_test("Test checks that error callback is called with correct code, when user denies access");
      setup({timeout: 20000});
      try {
        window.navigator.geolocation.getCurrentPosition(unexpectedSuccessCallback, errorCallback);
      } catch (err) {
        t.step(function () {
          assert_unreached("Error message: " + err.message);
        });
        t.done();
      }
      function unexpectedSuccessCallback (position) {
        t.step(function () {
          assert_unreached("successCallback should not be fired");
        });
        t.done();
      }
      function errorCallback (error) {
        t.step(function () {
          assert_true(error.code == error.PERMISSION_DENIED || error.code == error.POSITION_UNAVAILABLE, "error callback is called with correct code, when user denies access");
        });
        t.done();
      }
    </script>
  </body>
</html>
